Introduce a way to wait for a tcp server to be up.
This is used to wait for the fake SMTP server and the test webserver
to be ready. We do this by trying to connect to it repeatedly until
success, or a timeout of 1 second.

The introduction of this function prompted for rudimentary error
handling in the fake SMTP server (textproto.ReadLine() blocked).